+++ src/main/java/org/openjfx/API2Semestre/report/ReportInterval.java
